+++ lib/ModelMixins/Cesium3dTilesMixin.ts
@@ -313,8 +313,9 @@ export default function Cesium3dTilesMixin<
           return "";
         }
 
+        // Escape single quotes, cast property value to number
         const property =
-          "${feature['" + filter.property.replace(/'/g, "\\'") + "']}";
+          "Number(${feature['" + filter.property.replace(/'/g, "\\'") + "']})";
         const min =
           isDefined(filter.minimumValue) &&
           isDefined(filter.minimumShown) &&
@@ -327,6 +328,7 @@ export default function Cesium3dTilesMixin<
           filter.maximumShown < filter.maximumValue
             ? property + " <= " + filter.maximumShown
             : "";
+
         return [min, max].filter(x => x.length > 0).join(" && ");
       });