+++ CHANGES.md
@@ -3,6 +3,9 @@ Change Log
 
 #### next release (8.1.17)
 
+- **Breaking changes**:
+  * Minimum node version is now 12 after upgrading node-sass dependency
+
 * Automatically cast property value to number in style expressions generated for 3d tiles filter.
 * Re-enable procedure and observable selectors for SOS items.
 * Fix broken "Ideal zoom" for TableMixin items.