+++ CHANGES.md
@@ -24,7 +24,7 @@ Change Log
 * Upgrade compiler target from es2018 to es2019
 * Fix default table style legends
 * Remove SOS defaults legend workaround
-* [The next improvement]
+
 #### 8.1.16
 
 * Added region mapping support for Commonwealth Electoral Divisions as at 2 August 2021 (AEC) as com_elb_name_2021.